added Apply Laws and moved Composed extensions to mtl exten
