+++ modules/core/arrow-core-data/src/main/kotlin/arrow/typeclasses/Apply.kt
@@ -17,6 +17,9 @@ import arrow.Kind
 interface Apply<F> : Functor<F> {
   fun <A, B> Kind<F, A>.ap(ff: Kind<F, (A) -> B>): Kind<F, B>
 
+  infix fun <A, B, C> Kind<F, (A) -> B>.compose(fb: Kind<F, (B) -> C>) =
+    map2(fb) { x -> { a: A -> x.b(x.a(a)) } }
+
   fun <A, B, Z> map(
     a: Kind<F, A>,
     b: Kind<F, B>,
